Detach the kernel driver
On Linux, the SAM-BA bootloader is claimed by the cdc_acm driver,
preventing libnxt to access it.

Try to detach the kernel driver, if it fails, try to use the device
anyway.

The usb_detach_kernel_driver_np function is only implemented on Linux,
it will return an error on other system, just ignore it.

Fix issue reported at:
https://code.google.com/archive/p/libnxt/issues/2